requirements.txt
numpy==1.22.0
pandas==1.3.5
requests==2.27.1
schedule==1.1.0
urllib3==1.26.7
xmltodict==0.12.0
regex
mysql-connector
